Physics and render updates (#20)
* fix: move speeds now mean something

* feat: change some random table settings

* feat: add collision filtering

* fix: add collision mask for walls

+++ RogueliteSurvivor/RogueliteSurvivor/Components/Damage.cs
@@ -8,7 +8,7 @@ namespace RogueliteSurvivor.Components
 {
     public struct Damage
     {
-        public int Amount { get; set; }
-        public int BaseAmount { get; set; }
+        public float Amount { get; set; }
+        public float BaseAmount { get; set; }
     }
 }
